test handle exception

+++ server/src/main/java/com/elca/internship/server/dao/ProjectDAO.java
@@ -38,7 +38,7 @@ public interface ProjectDAO {
 
     void deleteByIds(List<Long> ids);
 
-    Long findProjectNumber(Long proNum);
+    Integer findProjectNumber(Integer proNum);
 
     int count();
